Adding code for resubmit

+++ src/main/java/net/floodlightcontroller/core/SwitchSyncRepresentation.java
@@ -136,7 +136,7 @@ public class SwitchSyncRepresentation {
     public SwitchSyncRepresentation(IOFSwitch sw) {
         this.dpid = sw.getId();
         this.buffers = sw.getBuffers();
-        this.tables = (short) sw.getNumTables();
+        this.tables = sw.getNumTables();
         this.capabilities = sw.getCapabilities();
         this.actions = sw.getActions();
         this.ports = toSyncedPortList(sw.getPorts());
